# backend/swagger.json
{
	"components": {
...
		"schemas": {
...
			"CategoryDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"categoryParentId": {
						"type": "string"
					},
					"status": {
						"type": "string",
						"enum": [
							"active",
							"inactive"
						]
					},
					"type": {
						"type": "string",
						"enum": [
							"news",
							"raovat",
							"menu"
						]
					},
					"styleShow": {
						"type": "string",
						"enum": [
							"news1",
							"news2",
							"news3"
						]
					},
					"createdAt": {
						"type": "string"
					},
					"slug": {
						"type": "string"
					},
					"parentCategory": {
						"properties": {
							"name": {
								"type": "string"
							},
							"id": {
								"type": "string"
							}
						},
						"type": "object"
					},
					"userNameUpdated": {
						"type": "string"
					},
					"updatedAt": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "string",
								"format": "date-time"
							}
						]
					}
				},
				"required": [
					"id",
					"name",
					"status",
					"type"
				],
				"type": "object",
				"additionalProperties": true
			},
...
			"PagedNewsItemDto": {
				"$ref": "#/components/schemas/Omit_NewsDto.content-or-updateAt-or-updatedBy-or-createdAt-or-createdBy_"
			},
			"PagedList_PagedNewsItemDto_": {
				"properties": {
					"items": {
						"items": {
							"$ref": "#/components/schemas/PagedNewsItemDto"
						},
						"type": "array"
					},
					"totalCount": {
						"type": "number",
						"format": "double"
					},
					"page": {
						"type": "number",
						"format": "double"
					},
					"size": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"items",
					"totalCount",
					"page",
					"size"
				],
				"type": "object",
				"additionalProperties": true
			},
			"NewStatus": {
				"enum": [
					"draft",
					"publish"
				],
				"type": "string"
			},
...
			"NewsDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"content": {
						"type": "string"
					},
					"publishDate": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"userId": {
						"type": "string"
					},
					"categoryId": {
						"type": "string"
					},
					"categoryName": {
						"type": "string"
					},
					"categoryStyle": {
						"type": "string"
					},
					"parentCategoryId": {
						"type": "string"
					},
					"parentCategoryName": {
						"type": "string"
					},
					"imageUrl": {
						"type": "string"
					},
					"status": {
						"type": "string",
						"enum": [
							"draft",
							"publish"
						]
					},
					"slug": {
						"type": "string"
					},
					"createdBy": {
						"type": "string"
					},
					"updatedBy": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"tagIds": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"isHotNews": {
						"type": "boolean"
					},
					"author": {
						"type": "string"
					},
					"userNameUpdated": {
						"type": "string"
					},
					"metaKeyword": {
						"type": "string"
					},
					"view": {
						"type": "string"
					},
					"index": {
						"type": "number",
						"format": "double"
					},
					"createdDate": {
						"type": "string",
						"format": "date-time"
					},
					"isPublish": {
						"type": "boolean"
					},
					"customId": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object",
				"additionalProperties": true
			},
			"NewsResponse": {
				"properties": {
					"data": {
						"$ref": "#/components/schemas/NewsDto"
					}
				},
				"required": [
					"data"
				],
				"type": "object",
				"additionalProperties": true
			},
			"NewsRequest": {
				"properties": {
					"id": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"content": {
						"type": "string"
					},
					"publishDate": {
						"type": "string",
						"format": "date-time"
					},
					"userId": {
						"type": "string"
					},
					"categoryId": {
						"type": "string"
					},
					"imageUrl": {
						"type": "string"
					},
					"status": {
						"type": "string",
						"enum": [
							"draft",
							"publish"
						]
					},
					"tagIds": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"isHotNews": {
						"type": "boolean"
					},
					"metaKeyword": {
						"type": "string"
					},
					"view": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": true
			},
			"RaoVatDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"content": {
						"type": "string"
					},
					"imageUrl": {
						"type": "string"
					},
					"extraImages": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"rawExtraImagePaths": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"categoryId": {
						"type": "string"
					},
					"categoryName": {
						"type": "string"
					},
					"publishDate": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "string",
								"format": "date-time"
							}
						]
					},
					"createdAt": {
						"type": "string"
					},
					"createdBy": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					},
					"updatedBy": {
						"type": "string"
					},
					"userNameUpdated": {
						"type": "string"
					},
					"slug": {
						"type": "string"
					},
					"facebook": {
						"type": "string"
					},
					"phoneNumber": {
						"type": "string"
					},
					"contactName": {
						"type": "string"
					},
					"metaKeyword": {
						"type": "string"
					},
					"websiteUrl": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"index": {
						"type": "number",
						"format": "double"
					},
					"view": {
						"type": "string"
					},
					"status": {
						"type": "string"
					},
					"customId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": true
			},
...
			"WebsiteDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"logo": {
						"type": "string"
					},
					"footerContent": {
						"type": "string"
					},
					"facebookUrl": {
						"type": "string"
					},
					"twitterUrl": {
						"type": "string"
					},
					"googleUrl": {
						"type": "string"
					},
					"linkedinUrl": {
						"type": "string"
					},
					"createdBy": {
						"type": "string"
					},
					"updatedBy": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"type": "object",
				"additionalProperties": true
			},
			"WebsiteResponse": {
				"properties": {
					"data": {
						"$ref": "#/components/schemas/WebsiteDto"
					}
				},
				"required": [
					"data"
				],
				"type": "object",
				"additionalProperties": true
			},
			"WebsiteRequest": {
				"properties": {
					"id": {
						"type": "string"
					},
					"logo": {
						"type": "string"
					},
					"footerContent": {
						"type": "string"
					},
					"facebookUrl": {
						"type": "string"
					},
					"twitterUrl": {
						"type": "string"
					},
					"googleUrl": {
						"type": "string"
					},
					"linkedinUrl": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": true
			},
...
			"AdvertisementPosition": {
				"type": "string",
				"enum": [
					"top",
					"left",
					"right"
				]
			},
...
			"AdvertisementDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"imageUrl": {
						"type": "string"
					},
					"url": {
						"type": "string"
					},
					"order": {
						"type": "number",
						"format": "double"
					},
					"position": {
						"$ref": "#/components/schemas/AdvertisementPosition"
					},
					"status": {
						"$ref": "#/components/schemas/AdvertisementStatus"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"userId": {
						"type": "string"
					},
					"userNameUpdated": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name",
					"imageUrl",
					"url",
					"order",
					"position",
					"status",
					"userId"
				],
				"type": "object",
				"additionalProperties": true
			},
...
			"TokenDto": {
				"properties": {
					"token": {
						"type": "string"
					},
					"refreshToken": {
						"type": "string"
					},
					"role": {
						"$ref": "#/components/schemas/RoleDto"
					},
					"userInfo": {
						"$ref": "#/components/schemas/UserDto"
					}
				},
				"required": [
					"token",
					"refreshToken"
				],
				"type": "object",
				"additionalProperties": true
			},
...
			"SignUpRequest": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"userType": {
						"$ref": "#/components/schemas/UserTypeEnum"
					}
				},
				"required": [
					"email",
					"password",
					"name",
					"userType"
				],
				"type": "object",
				"additionalProperties": true
			},
			"AccountCredentialsDto": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"isRememberMe": {
						"type": "boolean"
					}
				},
				"required": [
					"email",
					"password",
					"isRememberMe"
				],
				"type": "object",
				"additionalProperties": true
			},
			"UserSecretTokenDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"token": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"token"
				],
				"type": "object",
				"additionalProperties": true
			},
			"RefreshTokenRequest": {
				"properties": {
					"token": {
						"type": "string"
					}
				},
				"required": [
					"token"
				],
				"type": "object",
				"additionalProperties": true
			},
...
			"PagedList_PagedRaoVatItemDto_": {
				"properties": {
					"items": {
						"items": {
							"$ref": "#/components/schemas/PagedRaoVatItemDto"
						},
						"type": "array"
					},
					"totalCount": {
						"type": "number",
						"format": "double"
					},
					"page": {
						"type": "number",
						"format": "double"
					},
					"size": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"items",
					"totalCount",
					"page",
					"size"
				],
				"type": "object",
				"additionalProperties": true
			},
...
			"CreateCommentRequest": {
				"properties": {
					"newsId": {
						"type": "string"
					},
					"raoVatId": {
						"type": "string"
					},
					"content": {
						"type": "string"
					},
					"anonymousEmail": {
						"type": "string"
					},
					"anonymousName": {
						"type": "string"
					}
				},
				"required": [
					"content"
				],
				"type": "object",
				"additionalProperties": true
			},
...
			"PagedList_PagedCommentItemDto_": {
				"properties": {
					"items": {
						"items": {
							"$ref": "#/components/schemas/PagedCommentItemDto"
						},
						"type": "array"
					},
					"totalCount": {
						"type": "number",
						"format": "double"
					},
					"page": {
						"type": "number",
						"format": "double"
					},
					"size": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"items",
					"totalCount",
					"page",
					"size"
				],
				"type": "object",
				"additionalProperties": true
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"name": "authorization",
				"type": "apiKey",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "nguoivietplus-api",
		"version": "0.0.0",
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/admin-console/news": {
			"get": {
				"operationId": "GetNews",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PagedList_PagedNewsItemDto_"
								}
							}
						}
					}
				},
				"tags": [
					"AdminConsole"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "size",
						"required": false,
						"schema": {
							"default": 30,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/NewStatus"
						}
					}
				]
			},
			"post": {
				"operationId": "UpsertNews",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NewsResponse"
								}
							}
						}
					}
				},
				"tags": [
					"AdminConsole"
				],
				"security": [
					{
						"": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/NewsRequest"
							}
						}
					}
				}
			}
		},
...
		"/admin-console/website": {
			"get": {
				"operationId": "GetWebsite",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WebsiteResponse"
								}
							}
						}
					}
				},
				"tags": [
					"AdminConsole"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "UpsertWebsite",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WebsiteResponse"
								}
							}
						}
					}
				},
				"tags": [
					"AdminConsole"
				],
				"security": [
					{
						"": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/WebsiteRequest"
							}
						}
					}
				}
			}
		},
...
		"/advertisements": {
			"get": {
				"operationId": "GetAdvertisements",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/AdvertisementDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "size",
						"required": false,
						"schema": {
							"default": 30,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/NewStatus"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateAdvertisement",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AdvertisementDto"
								}
							}
						}
					}
				},
				"security": [
					{
						"": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"image": {
										"type": "string",
										"format": "binary"
									}
								},
								"required": [
									"image"
								]
							}
						}
					}
				}
			}
		},
		"/advertisements/{id}": {
			"get": {
				"operationId": "GetAdvertisement",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AdvertisementDto"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateAdvertisement",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AdvertisementDto"
								}
							}
						}
					}
				},
				"security": [
					{
						"": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": false,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"image": {
										"type": "string",
										"format": "binary"
									}
								}
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteAdvertisement",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [
					{
						"": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/auth/sign-up": {
			"post": {
				"operationId": "SignUp",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TokenDto"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SignUpRequest"
							}
						}
					}
				}
			}
		},
		"/auth/sign-in": {
			"post": {
				"operationId": "SignIn",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TokenDto"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AccountCredentialsDto"
							}
						}
					}
				}
			}
		},
		"/auth/sign-in-with-token": {
			"post": {
				"operationId": "SignInWithToken",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"userInfo": {
											"properties": {
												"email": {
													"type": "string"
												},
												"name": {
													"type": "string"
												},
												"id": {
													"type": "string"
												}
											},
											"required": [
												"email",
												"name",
												"id"
											],
											"type": "object"
										},
										"role": {
											"properties": {
												"name": {
													"type": "string"
												},
												"id": {
													"type": "string"
												}
											},
											"required": [
												"name",
												"id"
											],
											"type": "object"
										},
										"refreshToken": {
											"type": "string"
										},
										"token": {
											"type": "string"
										}
									},
									"required": [
										"userInfo",
										"role",
										"refreshToken",
										"token"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserSecretTokenDto"
							}
						}
					}
				}
			}
		},
		"/auth/refresh": {
			"post": {
				"operationId": "Refresh",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"userInfo": {
											"properties": {
												"email": {
													"type": "string"
												},
												"name": {
													"type": "string"
												},
												"id": {
													"type": "string"
												}
											},
											"required": [
												"email",
												"name",
												"id"
											],
											"type": "object"
										},
										"role": {
											"properties": {
												"name": {
													"type": "string"
												},
												"id": {
													"type": "string"
												}
											},
											"required": [
												"name",
												"id"
											],
											"type": "object"
										},
										"refreshToken": {
											"type": "string"
										},
										"token": {
											"type": "string"
										}
									},
									"required": [
										"userInfo",
										"role",
										"refreshToken",
										"token"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RefreshTokenRequest"
							}
						}
					}
				}
			}
		},
		"/auth/revoke": {
			"post": {
				"operationId": "Revoke",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [
					{
						"": []
					}
				],
				"parameters": []
			}
		},
...
		"/news": {
			"get": {
				"operationId": "GetNews",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PagedList_PagedNewsItemDto_"
								}
							}
						}
					}
				},
				"tags": [
					"website"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "size",
						"required": false,
						"schema": {
							"default": 30,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/NewStatus"
						}
					},
					{
						"in": "query",
						"name": "style",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"news1",
								"news2",
								"news3"
							]
						}
					},
					{
						"in": "query",
						"name": "isHotNews",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "menuSlug",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/rao-vat": {
			"get": {
				"operationId": "GetRaoVats",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PagedList_PagedRaoVatItemDto_"
								}
							}
						}
					}
				},
				"tags": [
					"website"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "size",
						"required": false,
						"schema": {
							"default": 30,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/NewStatus"
						}
					},
					{
						"in": "query",
						"name": "categorySlug",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/rao-vat/{slug}": {
			"get": {
				"operationId": "GetRaoVatBySlug",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RaoVatDto"
								}
							}
						}
					}
				},
				"tags": [
					"website"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "slug",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/categories": {
			"get": {
				"operationId": "GetCategoriesByType",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/CategoryDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"website"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "type",
						"required": true,
						"schema": {
							"type": "string",
							"enum": [
								"raovat",
								"menu",
								"news"
							]
						}
					}
				]
			}
		},
...
		"/comments": {
			"post": {
				"operationId": "AddComment",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"website"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateCommentRequest"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetComments",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PagedList_PagedCommentItemDto_"
								}
							}
						}
					}
				},
				"tags": [
					"website"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "size",
						"required": false,
						"schema": {
							"default": 30,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/NewStatus"
						}
					},
					{
						"in": "query",
						"name": "raoVatId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "newsId",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "http://localhost:3000/api"
		}
	]
}